Enum roqoqo_qiskit_devices::devices::IBMDevice

pub enum IBMDevice {
    IBMLagosDevice(IBMLagosDevice),
    IBMNairobiDevice(IBMNairobiDevice),
    IBMPerthDevice(IBMPerthDevice),
    IBMBelemDevice(IBMBelemDevice),
    IBMJakartaDevice(IBMJakartaDevice),
    IBMLimaDevice(IBMLimaDevice),
    IBMManilaDevice(IBMManilaDevice),
    IBMQuitoDevice(IBMQuitoDevice),
}

Collection of IBM quantum devices.

Variants

IBMLagosDevice(IBMLagosDevice)

IBMNairobiDevice(IBMNairobiDevice)

IBMPerthDevice(IBMPerthDevice)

IBMBelemDevice(IBMBelemDevice)

IBMJakartaDevice(IBMJakartaDevice)

IBMLimaDevice(IBMLimaDevice)

IBMManilaDevice(IBMManilaDevice)

IBMQuitoDevice(IBMQuitoDevice)

Implementations

impl IBMDevice

pub fn name(self) -> &'static str

Returns the IBM’s identifier.

Returns

A str of the name IBM uses as identifier.

pub fn set_single_qubit_gate_time( &mut self, gate: &str, qubit: usize, gate_time: f64 ) -> Result<(), RoqoqoError>

Setting the gate time of a single qubit gate.

Arguments

gate - hqslang name of the single-qubit-gate.

qubit - The qubit for which the gate time is set.

gate_time - gate time for the given gate.

pub fn set_two_qubit_gate_time( &mut self, gate: &str, control: usize, target: usize, gate_time: f64 ) -> Result<(), RoqoqoError>

Setting the gate time of a two qubit gate.

Arguments

gate - hqslang name of the two-qubit-gate.

control - The control qubit for which the gate time is set.

target - The target qubit for which the gate time is set.

gate_time - gate time for the given gate.

pub fn add_damping( &mut self, qubit: usize, damping: f64 ) -> Result<(), RoqoqoError>

Adds qubit damping to noise rates.

Arguments

qubit - The qubit for which the dampins is added.

daming - The damping rates.

pub fn add_dephasing( &mut self, qubit: usize, dephasing: f64 ) -> Result<(), RoqoqoError>

Adds qubit dephasing to noise rates.

Arguments

qubit - The qubit for which the dephasing is added.

dephasing - The dephasing rates.

pub fn to_generic_device(&self) -> Result<GenericDevice, RoqoqoError>

Converts the device to a qoqo GenericDevice.

Returns

GenericDevice - The converted device.

RoqoqoError - The error propagated from adding gate times and decoherence rates.